fixed simple temporal network class
