--- temporal_networks/temporal_network.py
@@ -1,631 +1,0 @@
-#from msilib.schema import Error
-from queue import PriorityQueue
-import json
-from xmlrpc.client import Boolean
-import copy
-from math import inf
-import numpy as np
-from scipy import stats
-import subprocess
-from graphviz import Digraph
-
-class SimpleTemporalNetwork:
-    """
-    represents a simple temporal network as a graph.
-    """
-    def __init__(self) -> None:
-        self.nodes : list[int] = []
-        self.labels : dict[int, str] = {}
-        self.edges = {}
-        self.edge_labels = {}
-
-    def add_node(self, node : int, label : str) -> None:
-        """
-        add a node to the network with label
-        """
-        self.nodes.append(node)
-        self.edges[node] = {node: 0}
-        self.edge_labels[node] = {node: "Self edge {}".format(node)}
-        self.labels[node] = label
-
-    def add_edge(self, node1 : int, node2 : int, label: str, duration_bound : float) -> None:
-        """
-        add an edge to the network with duration bound.
-        update the duration bound only if tighter than the current one.
-        """
-        print("\n", node1, node2, label)
-        if node1 not in self.edges:
-            self.edges[node1] = {}
-            self.edge_labels[node1] = {}
-        if node2 not in self.edges[node1]:
-            self.edges[node1][node2] = duration_bound
-            print(self.edge_labels)
-            print(self.edges)
-            self.edge_labels[node1][node2] = label
-        elif duration_bound < self.edges[node1][node2]:
-            self.edges[node1][node2] = duration_bound
-            self.edge_labels[node1][node2] = label
-
-    def floyd_warshall(self) -> bool:
-        """
-        use Floyd-Warshall to put the graph in all-pairs shortest path form.
-        returns True if the network is temporally consistent (no negative cycles)
-        """
-        # initialize missing edges 
-        for node1 in self.nodes:
-            for node2 in self.nodes:
-                if node1 != node2 and node2 not in self.edges[node1]:
-                    self.edges[node1][node2] = float("inf")
-        # run Floyd-Warshall
-        for k in self.nodes:
-            for i in self.nodes:
-                for j in self.nodes:
-                    self.edges[i][j] = min(self.edges[i][j], self.edges[i][k] + self.edges[k][j])
-                    # check for negative cycles
-                    if i==j and self.edges[i][j] < 0:
-                        return False
-        return True
-
-    def find_shortest_path(self, source : int, sink : int) -> float:
-        """
-        find the shortest path using dijkstras search
-        """
-        distances = dict.fromkeys(self.nodes, float("inf"))
-        distances[source] = 0
-        queue = PriorityQueue()
-        visited = set()
-        queue.put((0, source))
-        while not queue.empty():
-            distance, node = queue.get()
-            if node == sink: return distance
-            if node in visited: continue
-            visited.add(node)
-            if node not in self.edges: continue
-            for neighbor in self.edges[node]:
-                if neighbor in visited: continue
-                if distances[node] + self.edges[node][neighbor] < distances[neighbor]:
-                    distances[neighbor] = distances[node] + self.edges[node][neighbor]
-                    queue.put((distances[neighbor], neighbor))
-        return float("inf")
-
-    def make_minimal(self):
-        """
-        removes redundant edges from the network, assuming that the
-        network is temporally consistent and already in all-pairs
-        shortest path form.
-        Reference:
-        Nicola Muscettola, Paul Morris, and Ioannis Tsamardinos;
-        "Reformulating Temporal Plans For Efficient Execution";
-        In Principles of Knowledge Representation and Reasoning (1998).
-        """
-        for k in self.nodes:
-            for i in self.nodes:
-                if i == k: continue
-                if k not in self.edges[i]: continue
-                for j in self.nodes:
-                    if i == j or j == k: continue
-                    if j not in self.edges[k]: continue
-                    if j not in self.edges[i]: continue
-                    if self.edges[i][j] < self.edges[i][k] + self.edges[k][j]: continue
-                    if self.edges[i][j] < 0 and self.edges[i][k] < 0:
-                        del self.edges[i][j]
-                    elif self.edges[i][j] >=0 and self.edges[k][j] >= 0:
-                        del self.edges[i][j]
-                        
-    def print_dot_graph(self):
-        """
-        print the graph in DOT format.
-        """
-        print("digraph G {")
-        # declare nodes
-        for node in self.nodes:
-            print("\t" + str(node) + " [label=\"" + self.labels[node] + "\"];")
-        # declare edges
-        for node1 in self.nodes:
-            for node2 in self.edges[node1]:
-                if node1 == node2: continue
-                if self.edges[node1][node2] == float("inf"): continue
-                print("\t{} -> {} [label=\"{}: [{}]\"];".format(node1, node2, self.edge_labels[node1][node2], self.edges[node1][node2]))
-        print("}")
-
-    def print_graph_as_json(self):
-        """
-        print the graph in JSON format.
-        """
-        print("{")
-        print("\t\"nodes\": [")
-        for node in self.nodes:
-            print("\t\t{\"id\": " + str(node) + ", \"label\": \"" + self.labels[node] + "\"},")
-        print("\t],")
-        print("\t\"edges\": [")
-        for node1 in self.nodes:
-            for node2 in self.edges[node1]:
-                if node1 == node2: continue
-                if self.edges[node1][node2] == float("inf"): continue
-                print("\t\t{\"source\": " + str(node1) + ", \"target\": " + str(node2) + ", \"label\": \"" + str(self.edges[node1][node2]) + "\"},")
-        print("\t]")
-        print("}")
-
-class TemporalNetwork:
-    """
-    represents a simple temporal network as a graph.
-    """
-    def __init__(self) -> None:
-        self.name = None
-        self.time_points : list[TimePoint] = []
-        self.constraints: list[Constraint] = []
-    
-    def parse_from_temporal_plan_network(self, temporal_plan_network: SimpleTemporalNetwork):
-        """
-        Parses from an instance of TemporalPlanNetwork as output from the temporal plan. Changes bidirectional edges to be uni-directional.
-        """
-        for node in temporal_plan_network.nodes:
-            self.add_time_point(TimePoint(node, temporal_plan_network.labels[node]))
-
-        for node1 in temporal_plan_network.edges:
-            for node2 in temporal_plan_network.edges[node1]:
-                print("\nNode1: {}, Node2: {}, Duration: {}".format(node1, node2, temporal_plan_network.edges[node1][node2]))
-                if node1 == node2: continue
-                elif temporal_plan_network.edges[node1][node2] > 0:
-                    print("Constraint is an upper bound, adding edge {} -> {}".format(node1, node2))
-                    edge = Constraint(self.get_timepoint_by_id(node1), self.get_timepoint_by_id(node2), temporal_plan_network.edge_labels[node1][node2], "stc", {"lb": 0.01, "ub": temporal_plan_network.edges[node1][node2]})
-                elif temporal_plan_network.edges[node1][node2] < 0:
-                    print("Constraint is a lower bound, adding edge {} -> {}".format(node2, node1))
-                    edge = Constraint(self.get_timepoint_by_id(node2), self.get_timepoint_by_id(node1), temporal_plan_network.edge_labels[node1][node2],"stc", {"lb": -temporal_plan_network.edges[node1][node2], "ub": inf})
-                elif temporal_plan_network.edges[node1][node2] == 0:
-                    print("Constraint duration = 0, adding as a lower bound on edge {} -> {}".format(node2, node1))
-                    edge = Constraint(self.get_timepoint_by_id(node2), self.get_timepoint_by_id(node1), temporal_plan_network.edge_labels[node1][node2],"stc", {"lb": temporal_plan_network.edges[node1][node2], "ub": inf})
-                self.add_constraint(edge)
-
-    def copy(self):
-        """
-        returns a copy of the temporal network.
-        """
-        tn = TemporalNetwork()
-        tn.name = self.name
-        tn.time_points = copy.deepcopy(self.time_points)
-        tn.constraints = copy.deepcopy(self.constraints)
-        return tn
-    
-    def make_pstn(self):
-        """
-        returns a copy of the network as a probabilistic temporal network.
-        """
-        pstn = ProbabilisticTemporalNetwork()
-        pstn.name = self.name
-        pstn.time_points = copy.deepcopy(self.time_points)
-        pstn.constraints = copy.deepcopy(self.constraints)
-        return pstn
-
-    def add_time_point(self, time_point: TimePoint) -> None:
-        """
-        add a time-point (node) to the network.
-        """
-        for t in self.time_points:
-            if t.id == time_point.id:
-                raise ValueError("Time-point already exists in network with that ID. Try changing ID of new time-point so that it is unique.")
-        self.time_points.append(time_point)
-    
-    def add_name(self, name: str) -> None:
-        """
-        adds a string name to the network.
-        """
-        self.name = name
-
-    def add_constraint(self, constraint: Constraint) -> None:
-        """
-        add an edge (constraint) to the network. only constraints of type 'stc' are permitted in Temporal Network.
-        if source and sink nodes not in the time_points set, it adds them
-        """
-        assert constraint.type == "stc", "Only time-points of the type 'stc' are permitted."
-        # Checks if there is already a constraint with those time-points, if not it adds
-        existing = self.get_constraint_by_timepoint(constraint.source, constraint.sink)
-        if existing == None:
-            self.constraints.append(constraint)
-            if constraint.source not in self.time_points:
-                self.add_time_point(constraint.source)
-            if constraint.sink not in self.time_points:
-                self.add_time_point(constraint.sink)
-        # If the source and sink time-points are the same way round in the new constraint versus existing
-        elif existing.source == constraint.source:
-            # Checks whether the new constraint has a tighter bound
-            if constraint.ub < existing.ub:
-                existing.duration_bound["ub"] = constraint.ub
-            elif constraint.lb > existing.lb:
-                existing.duration_bound["lb"] = constraint.lb
-        # If the source and sink time-points are the wrong way round in the new constraint versus existing
-        elif existing.sink == constraint.source:
-            if -constraint.lb < existing.ub:
-                existing.duration_bound["ub"] = -constraint.lb
-            if -constraint.ub > existing.lb:
-                existing.duration_bound["lb"] = -constraint.ub
-
-    def get_adjacency_matrix(self) -> dict[TimePoint, dict]:
-        """
-        gets adjacency matrix (dictionary) representation of temporal-network
-        """
-        adj = {}
-        # Initialises adj matrix using edges explicit in self.constraints
-        for constraint in self.constraints:
-            # If source not in adj matrix yet, add it
-            if constraint.source.id not in adj:
-                adj[constraint.source.id] = {}
-            # If sink not in adj matrix yet, add it
-            if constraint.sink.id not in adj:
-                adj[constraint.sink.id] = {}
-            # If source[sink] not in adj matrix yet, add it.
-            if constraint.sink.id not in adj[constraint.source.id]:
-                adj[constraint.source.id][constraint.sink.id] = constraint.ub
-            # If sink[source] not in adj matrix yet, add it
-            if constraint.source.id not in adj[constraint.sink.id]:
-                adj[constraint.sink.id][constraint.source.id] = -constraint.lb
-        
-        # Adds self edges to be equal to zero and initialises missing edges to be infinity
-        for node1 in self.time_points:
-            if node1.id not in adj:
-                adj[node1.id] = {}
-            for node2 in self.time_points:
-                if node1 == node2:
-                    adj[node1.id][node2.id] = 0
-                elif node2.id not in adj[node1.id]:
-                    adj[node1.id][node2.id] = inf
-        return adj
-    
-    def get_bidirectional_network(self) -> dict[TimePoint, dict]:
-        """
-        Gets the bidirectional version of the temporal network, i.e. converts from l12 <= b2 - b1 <= u12 to b2 - b1 <= u12, b1 - b2 <= -l12
-        As above but does not consider all pairs.
-        """
-        network = {}
-        for constraint in self.constraints:
-            # If source not in adj matrix yet, add it
-            if constraint.source.id not in network:
-                network[constraint.source.id] = {}
-            # If sink not in adj matrix yet, add it
-            if constraint.sink.id not in network:
-                network[constraint.sink.id] = {}
-            # If source[sink] not in adj matrix yet, add it.
-            if constraint.sink.id not in network[constraint.source.id]:
-                network[constraint.source.id][constraint.sink.id] = constraint.ub
-            # If sink[source] not in adj matrix yet, add it
-            if constraint.source.id not in network[constraint.sink.id]:
-                network[constraint.sink.id][constraint.source.id] = -constraint.lb
-        return network
-
-
-    def floyd_warshall(self) -> tuple[dict[TimePoint, dict], bool]:
-        """
-        use Floyd-Warshall to put the graph in all-pairs shortest path form.
-        returns tuple (APSP dictionary, boolean) where the boolean is True if the network is consistent (i.e. no negative cycles).
-        """
-        adj = self.get_adjacency_matrix()
-
-        # run Floyd-Warshall
-        for k in adj:
-            for i in adj:
-                for j in adj:
-                    adj[i][j] = min(adj[i][j], adj[i][k] + adj[k][j])
-                    # check for negative cycles
-                    if i==j and adj[i][j] < 0:
-                        return (adj, False)
-        return (adj, True)
-
-    def find_shortest_path(self, source : int, sink : int) -> float:
-        """
-        find the shortest path using dijkstras search
-        """
-        # Needs updated
-        network = self.get_bidirectional_network()
-        distances = dict.fromkeys([i.id for i in self.time_points], float("inf"))
-        distances[source.id] = 0
-        queue = PriorityQueue()
-        visited = set()
-        queue.put((0, source))
-        while not queue.empty():
-            distance, node = queue.get()
-            if node == sink: return distance
-            if node in visited: continue
-            visited.add(node)
-            if node not in network: continue
-            for neighbor in network[node]:
-                if neighbor in visited: continue
-                if distances[node] + network[node][neighbor] < distances[neighbor]:
-                    distances[neighbor] = distances[node] + network[node][neighbor]
-                    queue.put((distances[neighbor], neighbor))
-        return float("inf")
-
-    def make_minimal(self) -> dict[TimePoint]:
-        """
-        removes redundant edges from the network, assuming that the
-        network is temporally consistent and already in all-pairs
-        shortest path form.
-        Reference:
-        Nicola Muscettola, Paul Morris, and Ioannis Tsamardinos;
-        "Reformulating Temporal Plans For Efficient Execution";
-        In Principles of Knowledge Representation and Reasoning (1998).
-        """
-        check = self.floyd_warshall()
-        adj, consistent = check[0], check[1]
-        if consistent == False:
-            raise AttributeError("Network is not consistent")    
-        for k in adj:
-            for i in adj:
-                if i == k: continue
-                if k not in adj[i]: continue
-                for j in adj:
-                    if i == j or j == k: continue
-                    if j not in adj[k]: continue
-                    if j not in adj[i]: continue
-                    if adj[i][j] < adj[i][k] + adj[k][j]: continue
-                    if adj[i][j] < 0 and adj[i][k] < 0:
-                        del adj[i][j]
-                    elif adj[i][j] >=0 and adj[k][j] >= 0:
-                        del adj[i][j]
-        return adj
-    
-    def get_outgoing_edge(self, constraint: Constraint) -> list[Constraint]:
-        """
-        given an edge (i, j), returns a list of outgoing edges (j, k)
-        """
-        return [jk for jk in self.constraints if jk.source == constraint.sink]
-    
-    def get_incoming_edge(self, constraint: Constraint) -> list[Constraint]:
-        """
-        given an edge (j, k), returns a list of incoming edges (i, j)
-        """
-        return [ij for ij in self.constraints if ij.sink == constraint.source]
-    
-    def get_constraint_by_timepoint(self, source: TimePoint, sink: TimePoint) -> Constraint:
-        """
-        given two time-points, i and j, if a constraint exists between the two it returns the constraint, else raises exception 
-        """
-        for constraint in self.constraints:
-            if constraint.source == source and constraint.sink == sink:
-                return constraint
-            elif constraint.sink == source and constraint.source == sink:
-                return constraint
-            else:
-                return None
-
-    def get_timepoint_by_id(self, id: int) -> TimePoint:
-        """
-        given an id, it returns the time-point if it exists in self.timepoints
-        """
-        found = None
-        for time_point in self.time_points:
-            if time_point.id == id:
-                found = time_point
-        return found
-            
-    def print_dot_graph(self):
-        """
-        print the graph in DOT format.
-        """
-        print("digraph G {")
-        # declare nodes
-        for time_point in self.time_points:
-            print("\t" + str(time_point.id) + " [label=\"" + time_point.label + "\"];")
-        # declare edges
-        for constraint in self.constraints:
-            print("\t{} -> {} [label=\"{}: [{}, {}]\"];".format(constraint.source.id, constraint.sink.id, constraint.label, constraint.lb, constraint.ub))
-        print("}")
-    
-    def plot_dot_graph(self):
-        """
-        Plot the graph as a dot graph using graphviz
-        """
-        plot = Digraph()
-        for timePoint in self.time_points:
-            plot.node(name=str(timePoint.id), label=str(timePoint.id))
-        
-        for constraint in self.constraints:
-            plot.edge(str(constraint.source.id), str(constraint.sink.id), label="{}: [{:.3f}, {:.3f}]".format(constraint.label, constraint.lb, constraint.ub))
-        try:
-            plot.render('logs/{}_plot.png'.format(self.name), view=True)
-        except subprocess.CalledProcessError:
-            print("Please close the PDF and rerun the script")
-
-    def print_graph_as_json(self):
-        """
-        print the graph in JSON format.
-        """
-        print("{")
-        print("\t\"timepoints\": [")
-        for time_point in self.time_points:
-            print("\t\t{\"id\": " + str(time_point.id) + ", \"label\": \"" + time_point.label + "\"},")
-        print("\t],")
-        print("\t\"constraints\": [")
-        for constraint in self.constraints:
-            if constraint.type == "stc":
-                print("\t\t{\"source\": " + str(constraint.source.id) + ", \"target\": " + str(constraint.sink.id) + ", \"label\": \"" + constraint.label + "\", \"bounds\": " + "({}, {})".format(constraint.lb, constraint.ub) + "},")
-            elif constraint.type == "pstc":
-                print("\t\t{\"source\": " + str(constraint.source.id) + ", \"target\": " + str(constraint.sink.id) + ", \"label\": \"" + constraint.label + "\", \"distribution\": " + "N({}, {})".format(constraint.mean, constraint.sd) + "},")
-        print("\t]")
-        print("}")
-    
-    def save_as_json(self, filename):
-        """
-        saves the network as a JSON to filename.json
-        """
-        toDump = {}
-        toDump["timepoints"] = [t.to_json() for t in self.time_points]
-        toDump["constraints"] = [c.to_json() for c in self.constraints]
-        with open("{}.json".format(filename), 'w') as fp:
-            json.dump(toDump, fp)
-
-class ProbabilisticTemporalNetwork(TemporalNetwork):
-    """
-    represents a probabilistic temporal network.
-    """
-    def __init__(self) -> None:
-        super().__init__()
-
-    def adds_constraint(self, constraint: Constraint) -> None:
-        """
-        add an edge (constraint) to the network. permits constraints of type 'pstc'.
-        """
-        # Checks if there is already a constraint with those time-points, if not it adds
-        existing = self.get_constraint_by_timepoint(constraint.source, constraint.sink)
-        if existing == None:
-            self.constraints.append(constraint)
-            if constraint.source not in self.time_points:
-                self.add_time_point(constraint.source)
-            if constraint.sink not in self.time_points:
-                self.add_time_point(constraint.sink)
-        # If the source and sink time-points are the same way round in the new constraint versus existing
-        elif existing.source == constraint.source:
-            # Checks whether the new constraint has a tighter bound
-            if constraint.ub < existing.ub:
-                existing.ub = constraint.ub
-            elif constraint.lb > existing.lb:
-                existing.lb = constraint.lb
-        # If the source and sink time-points are the wrong way round in the new constraint versus existing
-        elif existing.sink == constraint.source:
-            if -constraint.lb < existing.ub:
-                existing.ub = -constraint.lb
-            if -constraint.ub > existing.lb:
-                existing.lb = -constraint.ub
-    
-    def get_probabilistic_constraints(self) -> list[Constraint]:
-        """
-        returns a list of probabilistic constraints (those with type = pstc)
-        """
-        return [i for i in self.constraints if i.type == "pstc"]
-
-    
-    def get_requirement_constraints(self) -> list[Constraint]:
-        """
-        returns a list of requirement constraints (those with type = stc)
-        """
-        return [i for i in self.constraints if i.type == "stc"]
-
-    def set_controllability_of_time_points(self) -> None:
-        """
-        checks which time_points are uncontrollable (i.e. they come at the end of a probabilistic constraint). Sets the controllable flag
-        to True if controllable and False if not controllable
-        """
-        uncontrollable_time_points = [i.sink for i in self.get_probabilistic_constraints()]
-        for time_point in self.time_points:
-            if time_point in uncontrollable_time_points:
-                time_point.controllable = False
-            else:
-                time_point.controllable = True
-    
-    def get_controllable_time_points(self) -> list[TimePoint]:
-        """
-        returns a list of controllable time-points
-        """
-        self.set_controllability_of_time_points()
-        return [i for i in self.time_points if i.controllable == True]
-    
-    def get_uncontrollable_time_points(self) -> list[TimePoint]:
-        """
-        returns a list of uncontrollable time-points
-        """
-        self.set_controllability_of_time_points()
-        return [i for i in self.time_points if i.controllable == False]
-    
-    def get_uncontrollable_constraints(self) -> list[Constraint]:
-        """
-        returns a list of requirement constraints that contain an uncontrollable time-point
-        """
-        self.set_controllability_of_time_points()
-        uncontrollable_constraints = []
-        for constraint in self.constraints:
-            if constraint.source.controllable == False or constraint.sink.controllable == False:
-                uncontrollable_constraints.append(constraint)
-        return uncontrollable_constraints
-    
-    def get_controllable_constraints(self) -> list[Constraint]:
-        self.set_controllability_of_time_points()
-        return [i for i in self.get_requirement_constraints() if i.source.controllable == True and i.sink.controllable == True]
-    
-    def incoming_probabilistic(self, constraint: Constraint) -> dict[str, Constraint]:
-        """
-        returns a dictionary of the incoming probabilistic constraint in the form {"start": Constraint, "end": Constraint}
-        raises an exception if the number of incoming probabilistic constraints is greater than one
-        """
-        if constraint not in self.get_uncontrollable_constraints():
-            return None
-        else:
-            incoming_source = [g for g in self.getContingents() if g.sink == constraint.source]
-            incoming_sink = [g for g in self.getContingents() if g.sink == constraint.sink]
-            if len(incoming_source) > 1 or len(incoming_sink) > 1:
-                raise AttributeError("More than one incoming probabilistic edge.")
-            else:
-                try:
-                    return {"start": incoming_source[0], "end": incoming_sink[0]}
-                except IndexError:
-                    try:
-                        return {"start": incoming_source[0], "end": None}
-                    except IndexError:
-                        return {"start": None, "end": incoming_sink[0]}
-
-    def read_uncertainties_from_json(self, file: json):
-        """
-        Reads in a json of action and til uncertainties, such that the uncertainty x = sd/mean. Updates edges with
-        distributions and makes edges probabilistic. Returns a PSTN.
-        """
-        with open(file) as f:
-            uncertainties = json.load(f)
-        actions, tils = uncertainties["actions"], uncertainties["tils"]
-
-        for action in actions:
-            for constraint in self.constraints:
-                if action["name"] in constraint.label:
-                    if constraint.type == "stc":
-                        assert constraint.ub == constraint.lb
-                        constraint.distribution = {"mean": constraint.ub, "sd": constraint.ub * action["uncertainty"]}
-                        constraint.type = "pstc"
-                    else:
-                        raise ValueError("Uncertainties already added to costraints.")
-        for til in tils:
-            for constraint in self.constraints:
-                if til["name"] in constraint.label:
-                    if constraint.type == "stc":
-                        assert constraint.ub == constraint.lb
-                        constraint.distribution = {"mean": constraint.ub, "sd": constraint.ub * til["uncertainty"]}
-                        constraint.type = "pstc"
-                    else:
-                        raise ValueError("Uncertainties already added to constraints.")
-
-    def make_correlated_pstn(self):
-        """
-        returns a copy of the network as a correlated probabilistic temporal network.
-        """
-        corr_pstn = CorrelatedProbabilisticTemporalNetwork()
-        corr_pstn.name = self.name
-        corr_pstn.time_points = copy.deepcopy(self.time_points)
-        corr_pstn.constraints = copy.deepcopy(self.constraints)
-        return corr_pstn
-    
-    def plot_dot_graph(self):
-        """
-        Plot the graph as a dot graph using graphviz
-        """
-        requirements = self.get_requirement_constraints()
-        probabilistics = self.get_probabilistic_constraints()
-
-        plot = Digraph()
-        for timePoint in self.time_points:
-            plot.node(name=str(timePoint.id), label=str(timePoint.id))
-        
-        for constraint in requirements:
-            plot.edge(str(constraint.source.id), str(constraint.sink.id), label="{}: [{}, {}]".format(constraint.label, constraint.lb, constraint.ub))
-        for constraint in probabilistics:
-            plot.edge(str(constraint.source.id), str(constraint.sink.id), label="{}: N({}, {})".format(constraint.label, constraint.mean, constraint.sd))
-        try:
-            plot.render('logs/{}.png'.format(self.name), view=True)
-        except subprocess.CalledProcessError:
-            print("Please close the PDF and rerun the script")
-    
-class CorrelatedProbabilisticTemporalNetwork(ProbabilisticTemporalNetwork):
-    """
-    represents a correlated probabilistic temporal network.
-    """
-    def __init__(self) -> None:
-        super().__init__()
-        self.correlations = []
-    
-    def add_correlation(self, constraints: list[Constraint]) -> None:
-        self.correlations.append(Correlation(constraints))
